enhance-init-master-node
