--- clients/.gitignore
@@ -1,2 +1,0 @@
-tsunami
-nohup.out
