--- clients/job.go
@@ -1,5 +1,0 @@
-package main
-
-// Job detial
-type Job struct {
-}
