+++ master/makefile
@@ -1,0 +1,10 @@
+all: ocean 
+
+clean:
+	rm ocean 
+
+ocean:
+	go build ocean.go services.go 
+
+install:
+	cp -f ocean /usr/local/bin
