+++ worker/.gitignore
@@ -1,0 +1,2 @@
+tsunami
+nohup.out
