+++ worker/Godeps/Godeps.json
@@ -1,0 +1,6 @@
+{
+	"ImportPath": "github.com/tsunami/clients",
+	"GoVersion": "go1.10",
+	"GodepVersion": "v80",
+	"Deps": []
+}
