+++ worker/job.go
@@ -1,0 +1,5 @@
+package main
+
+// Job detial
+type Job struct {
+}
