+++ worker/makefile
@@ -1,0 +1,10 @@
+all: tsunami
+
+clean:
+	rm tsunami
+
+tsunami:
+	go build tsunami.go json.go shell.go help.go config.go job.go worker.go utils.go
+
+install:
+	cp -f tsunami /usr/local/bin
