+++ worker/utils.go
@@ -1,0 +1,19 @@
+package main
+
+import (
+	"log"
+	"net"
+)
+
+// GetIP get local ip
+func GetIP() net.IP {
+	conn, err := net.Dial("udp", "8.8.8.8:80")
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer conn.Close()
+
+	localAddr := conn.LocalAddr().(*net.UDPAddr)
+
+	return localAddr.IP
+}
